feat(algo/optim/lbfgs): implement L-BFGS algorithm
