+++ include/nuri/algo/optim.h
@@ -21,21 +21,19 @@
 #include "nuri/utils.h"
 
 namespace nuri {
-enum class LbfgsbResultCode {
+enum class LbfgsResultCode {
   kSuccess,
   kMaxIterReached,
   kInvalidInput,
   kAbnormalTerm,
 };
 
-struct LbfgsbResult {
-  LbfgsbResultCode code;
+struct LbfgsResult {
+  LbfgsResultCode code;
   int niter;
   double fx;
   ArrayXd gx;
 };
-
-class LBfgsB;
 
 namespace internal {
   constexpr double kEpsMach = 2.220446049250313e-16;
@@ -50,8 +48,10 @@ namespace internal {
 
   class Dcsrch {
   public:
-    Dcsrch(double f0, double g0, double step0, double stepmin = 0,
-           double stepmax = 1e+10, double ftol = 1e-3, double gtol = 0.9,
+    constexpr static double kStepMin = 0, kStepMax = 1e+10;
+
+    Dcsrch(double f0, double g0, double step0, double stepmin = kStepMin,
+           double stepmax = kStepMax, double ftol = 1e-3, double gtol = 0.9,
            double xtol = 0.1) noexcept;
 
     DcsrchStatus operator()(double f, double g);
@@ -85,6 +85,278 @@ namespace internal {
     bool brackt_ = false, bisect_ = false;
   };
 
+  class LBfgsBase {
+  public:
+    LBfgsBase(MutRef<ArrayXd> x, int m);
+
+    /* State modifiers, only for implementations */
+
+    int n() const { return static_cast<int>(x_.size()); }
+
+    int m() const { return static_cast<int>(ws_.cols()); }
+
+    auto &x() { return x_; }
+
+    const auto &x() const { return x_; }
+
+    auto wnt() { return wnt_.topLeftCorner(2 * col_, 2 * col_); }
+    auto &wnt_raw() { return wnt_; }
+
+    auto ws() { return ws_.leftCols(col_); }
+
+    auto wy() { return wy_.leftCols(col_); }
+
+    auto ss() { return ss_.topLeftCorner(col_, col_); }
+
+    auto sy() { return sy_.topLeftCorner(col_, col_); }
+
+    auto wtt() { return wtt_.topLeftCorner(col_, col_); }
+
+    auto p() { return p_.head(2 * col_); }
+
+    auto v() { return v_.head(2 * col_); }
+
+    auto c() { return c_.head(2 * col_); }
+
+    auto wbp() { return wbp_.head(2 * col_); }
+
+    auto smul() { return smul_.head(nonnegative(col_ - 1)); }
+
+    auto &z() { return z_; }
+    auto &xcp() { return z_; }
+
+    auto &r() { return r_; }
+
+    auto &t() { return t_; }
+    auto &xp() { return t_; }
+
+    auto &d() { return d_; }
+
+    double d_dot(const ArrayXd &gx) const {
+      return gx.matrix().dot(d_.matrix());
+    }
+
+    double theta() const { return theta_; }
+
+    int col() const { return col_; }
+
+    int prev_col() const { return prev_col_; }
+
+    bool updated() const { return updated_; }
+
+    /* Test utils; normally not used */
+
+    void update_col(int col) { col_ = col; }
+
+    void update_theta(double theta) { theta_ = theta; }
+
+  protected:
+    bool prepare_next_iter(double gd, double ginit, double step, double dtd);
+
+    void reset_memory() {
+      theta_ = 1.0;
+      col_ = prev_col_ = 0;
+      updated_ = false;
+    }
+
+  private:
+    /*  System, size n */
+    MutRef<ArrayXd> x_;
+
+    /* (2m, 2m) */
+    MatrixXd wnt_;
+    /* (n, m) */
+    MatrixXd ws_, wy_;
+    /* (m, m) */
+    MatrixXd ss_, sy_, wtt_;
+    /* 2*m */
+    VectorXd p_, v_, c_, wbp_;
+    /* n */
+    ArrayXd z_, r_, t_, d_;
+    /* m-1 */
+    ArrayXd smul_;
+
+    double theta_;
+    int col_, prev_col_;
+    bool updated_;
+  };
+
+  extern bool lbfgs_bmv(MutVecBlock<VectorXd> p, MutVecBlock<ArrayXd> smul,
+                        ConstRef<VectorXd> v, ConstRef<MatrixXd> sy,
+                        ConstRef<MatrixXd> wtt);
+}  // namespace internal
+
+/**
+ * @brief L-BFGS(-B) minimizer
+ * @sa l_bfgs_b, l_bfgs
+ *
+ * References:
+ * - RH Byrd, P Lu, J Nocedal and C Zhu. *SIAM J. Sci. Stat. Comput.* **1995**,
+ *   *16* (5), 1190-1208. DOI:[10.1137/0916069](https://doi.org/10.1137/0916069)
+ * - C Zhu, RH Byrd, J Nocedal. *ACM Trans. Math. Softw.* **1997**, *23* (4),
+ *   550-560. DOI:[10.1145/279232.279236](https://doi.org/10.1145/279232.279236)
+ * - JL Morales, J Nocedal. *ACM Trans. Math. Softw.* **2011**, *38* (1), 7.
+ *   DOI:[10.1145/2049662.2049669](https://doi.org/10.1145/2049662.2049669)
+ *
+ * This implementation is based on the C implementation of L-BFGS-B in the SciPy
+ * library, which is a translation of the original Fortran code by Ciyou Zhu,
+ * Richard Byrd, and Jorge Noceda. Both are released under the BSD 3-Clause
+ * License and the original license is included below.
+ *
+ * \code{.unparsed}
+ * Copyright (c) 2011 Ciyou Zhu, Richard Byrd, Jorge Nocedal and Jose Luis
+ * Morales.
+ * Copyright (c) 2001-2002 Enthought, Inc. 2003-2024, SciPy Developers.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above
+ *    copyright notice, this list of conditions and the following
+ *    disclaimer in the documentation and/or other materials provided
+ *    with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * \endcode
+ */
+template <class Impl>
+class LBfgs: public internal::LBfgsBase {
+public:
+  /**
+   * @brief Prepare L-BFGS(-B) minimization algorithm.
+   *
+   * @param x Initial guess. Will be modified in-place.
+   * @param impl Implementation object.
+   * @param m The maximum number of variable metric corrections used to define
+   *        the limited memory matrix.
+   */
+  LBfgs(MutRef<ArrayXd> x, Impl &&impl, const int m = 10)
+      : internal::LBfgsBase(x, m), impl_(std::move(impl)) { }
+
+  /**
+   * @brief Minimize a function using L-BFGS(-B) algorithm.
+   *
+   * @tparam FuncGrad Function object that computes the function value and
+   *         gradient. Function value should be returned and gradient should be
+   *         updated in the input gradient vector.
+   * @param fg Function object.
+   * @param factr Stop when function value changes by less than this factor
+   *        times the machine precision.
+   * @param pgtol Stop when the projected gradient is less than this value.
+   * @param maxiter Maximum number of iterations.
+   * @param maxls Maximum number of line search steps.
+   * @return A struct with the result code, number of iterations, final function
+   *         value, and final gradient.
+   */
+  template <class FuncGrad>
+  LbfgsResult minimize(FuncGrad fg, double factr = 1e+7, double pgtol = 1e-5,
+                       int maxiter = 15000, int maxls = 20);
+
+  Impl &impl() { return impl_; }
+
+private:
+  Impl impl_;
+};
+
+template <class Impl>
+template <class FuncGrad>
+LbfgsResult LBfgs<Impl>::minimize(FuncGrad fg, const double factr,
+                                  const double pgtol, const int maxiter,
+                                  const int maxls) {
+  const double tol = factr * internal::kEpsMach;
+
+  ArrayXd gx(n());
+  double fx = fg(gx, x());
+
+  internal::AllowEigenMallocScoped<false> ems;
+
+  double sbgnrm = impl_.projgr(x(), gx);
+  if (sbgnrm <= pgtol)
+    return { LbfgsResultCode::kSuccess, 0, fx, std::move(gx) };
+
+  impl_.reset();
+  reset_memory();
+
+  int iter = 0;
+  for (; iter < maxiter; ++iter) {
+    if (!impl_.prepare_lnsrch(*this, gx, sbgnrm, iter)) {
+      reset_memory();
+      continue;
+    }
+
+    d() = z() - x();
+
+    double gd = d_dot(gx);
+    if (gd >= 0) {
+      if (col() == 0)
+        return { LbfgsResultCode::kAbnormalTerm, iter + 1, fx, std::move(gx) };
+
+      reset_memory();
+      continue;
+    }
+
+    t() = x();
+    r() = gx;
+
+    auto lnsrl = impl_.lnsrch(x(), t(), z(), d(), fx, gd, iter);
+    bool converged = false;
+    for (int i = 0; !converged && i < maxls; ++i) {
+      NURI_EIGEN_ALLOW_MALLOC(true);
+      fx = fg(gx, x());
+      NURI_EIGEN_ALLOW_MALLOC(false);
+
+      gd = d_dot(gx);
+      converged = lnsrl.search(fx, gd);
+    }
+
+    if (!converged) {
+      x() = t();
+      fx = lnsrl.finit();
+      gx = r();
+
+      if (col() == 0)
+        return { LbfgsResultCode::kAbnormalTerm, iter + 1, fx, std::move(gx) };
+
+      reset_memory();
+      continue;
+    }
+
+    sbgnrm = impl_.projgr(x(), gx);
+    if (sbgnrm <= pgtol)
+      return { LbfgsResultCode::kSuccess, iter + 1, fx, std::move(gx) };
+
+    double ddum = std::max({ std::abs(lnsrl.finit()), std::abs(fx), 1.0 });
+    if (lnsrl.finit() - fx <= tol * ddum)
+      return { LbfgsResultCode::kSuccess, iter + 1, fx, std::move(gx) };
+
+    r() = gx - r();
+    if (!prepare_next_iter(gd, lnsrl.ginit(), lnsrl.step(), lnsrl.dtd()))
+      reset_memory();
+  }
+
+  return { LbfgsResultCode::kMaxIterReached, maxiter, fx, std::move(gx) };
+}
+
+namespace internal {
   /**
    * nbd == 0x1 if has lower bound,
    *        0x2 if has upper bound,
@@ -155,8 +427,6 @@ namespace internal {
 
     double ginit() const { return dcsrch_.ginit(); }
 
-    constexpr static double kStepMin = 0, kStepMax = 1e+10;
-
   private:
     MutRef<ArrayXd> &x() { return *x_; }
 
@@ -187,312 +457,78 @@ namespace internal {
     return lhs.tj > rhs.tj;
   }
 
+  class LBfgsBImpl {
+  public:
+    LBfgsBImpl(MutRef<ArrayXd> x, LbfgsbBounds bounds, int m);
+
+    LbfgsbLnsrch lnsrch(MutRef<ArrayXd> &x, const ArrayXd &t, const ArrayXd &z,
+                        const ArrayXd &d, double f0, double g0, int iter,
+                        double ftol = 1e-3, double gtol = 0.9,
+                        double xtol = 0.1) const noexcept;
+
+    void reset();
+
+    bool prepare_lnsrch(LBfgs<LBfgsBImpl> &lbfgsb, const ArrayXd &gx,
+                        double sbgnrm, int iter);
+
+    double projgr(ConstRef<ArrayXd> x, const ArrayXd &gx);
+
+    // For internal use only
+
+    MatrixXd &wn1() { return wn1_; }
+
+    int n() const { return static_cast<int>(iwhere_.size()); }
+
+    const LbfgsbBounds &bounds() const { return bounds_; }
+
+    int nfree() const { return nfree_; }
+
+    bool constrained() const { return constrained_; }
+
+    bool boxed() const { return boxed_; }
+
+    auto &iwhere() { return iwhere_; }
+
+    auto free() const { return free_bound_.head(nfree_); }
+
+    auto bound() const { return free_bound_.tail(n() - nfree_); }
+
+    auto enter() const { return enter_leave_.head(nenter_); }
+
+    auto leave() const { return enter_leave_.tail(nleave_); }
+
+    auto &brks() { return brks_; }
+
+    ArrayXi &free_bound() { return free_bound_; }
+
+    void update_nfree(int nfree) { nfree_ = nfree; }
+
+  private:
+    bool freev(int iter);
+
+    /* (2m, 2m) */
+    MatrixXd wn1_;
+
+    /* Bound constraints */
+    LbfgsbBounds bounds_;
+
+    /* n */
+    ArrayXi iwhere_, free_bound_, enter_leave_;
+
+    ClearablePQ<CauchyBrkpt, std::greater<>> brks_;
+
+    int nfree_, nenter_, nleave_;
+    bool constrained_, boxed_;
+  };
+
   extern bool lbfgsb_errclb(const ArrayXd &x, const ArrayXi &nbd,
                             const Array2Xd &bounds, int m, double factr);
 
-  extern double lbfgsb_projgr(ConstRef<ArrayXd> x, const ArrayXd &gx,
-                              const LbfgsbBounds &bounds);
+  extern bool lbfgsb_cauchy(LBfgs<LBfgsBImpl> &lbfgsb, const ArrayXd &gx,
+                            double sbgnrm);
 
-  extern bool lbfgs_bmv(MutVecBlock<VectorXd> p, MutVecBlock<ArrayXd> smul,
-                        ConstRef<VectorXd> v, ConstRef<MatrixXd> sy,
-                        ConstRef<MatrixXd> wtt);
-
-  extern bool lbfgsb_cauchy(LBfgsB &lbfgsb, const ArrayXd &gx, double sbgnrm);
-
-  extern bool lbfgsb_subsm(LBfgsB &lbfgsb, const ArrayXd &gg);
+  extern bool lbfgsb_subsm(LBfgs<LBfgsBImpl> &lbfgsb, const ArrayXd &gg);
 }  // namespace internal
-
-/**
- * @brief L-BFGS-B minimizer
- * @sa l_bfgs_b
- *
- * References:
- * - RH Byrd, P Lu, J Nocedal and C Zhu. *SIAM J. Sci. Stat. Comput.* **1995**,
- *   *16* (5), 1190-1208. DOI:[10.1137/0916069](https://doi.org/10.1137/0916069)
- * - C Zhu, RH Byrd, J Nocedal. *ACM Trans. Math. Softw.* **1997**, *23* (4),
- *   550-560. DOI:[10.1145/279232.279236](https://doi.org/10.1145/279232.279236)
- * - JL Morales, J Nocedal. *ACM Trans. Math. Softw.* **2011**, *38* (1), 7.
- *   DOI:[10.1145/2049662.2049669](https://doi.org/10.1145/2049662.2049669)
- *
- * This implementation is based on the C implementation of L-BFGS-B in the SciPy
- * library, which is a translation of the original Fortran code by Ciyou Zhu,
- * Richard Byrd, and Jorge Noceda. Both are released under the BSD 3-Clause
- * License and the original license is included below.
- *
- * \code{.unparsed}
- * Copyright (c) 2011 Ciyou Zhu, Richard Byrd, Jorge Nocedal and Jose Luis
- * Morales.
- * Copyright (c) 2001-2002 Enthought, Inc. 2003-2024, SciPy Developers.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above
- *    copyright notice, this list of conditions and the following
- *    disclaimer in the documentation and/or other materials provided
- *    with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- * \endcode
- */
-class LBfgsB {
-public:
-  /**
-   * @brief Prepare L-BFGS-B minimization algorithm.
-   *
-   * @param x Initial guess. Will be modified in-place.
-   * @param bounds Bounds for each variable.
-   * @param m The maximum number of variable metric corrections used to define
-   *        the limited memory matrix.
-   */
-  LBfgsB(MutRef<ArrayXd> x, internal::LbfgsbBounds bounds, int m = 10);
-
-  /**
-   * @brief Minimize a function using L-BFGS-B algorithm.
-   *
-   * @tparam FuncGrad Function object that computes the function value and
-   *         gradient. Function value should be returned and gradient should be
-   *         updated in the input gradient vector.
-   * @param fg Function object.
-   * @param factr Stop when function value changes by less than this factor
-   *        times the machine precision.
-   * @param pgtol Stop when the projected gradient is less than this value.
-   * @param maxiter Maximum number of iterations.
-   * @param maxls Maximum number of line search steps.
-   * @return A struct with the result code, number of iterations, final function
-   *         value, and final gradient.
-   */
-  template <class FuncGrad>
-  LbfgsbResult minimize(FuncGrad fg, double factr = 1e+7, double pgtol = 1e-5,
-                        int maxiter = 15000, int maxls = 20);
-
-  /* State modifiers, only for implementations */
-
-  int n() const { return static_cast<int>(x_.size()); }
-
-  int m() const { return static_cast<int>(ws_.cols()); }
-
-  auto &x() { return x_; }
-
-  const auto &x() const { return x_; }
-
-  const internal::LbfgsbBounds &bounds() const { return bounds_; }
-
-  bool constrained() const { return constrained_; }
-
-  bool boxed() const { return boxed_; }
-
-  auto wnt() { return wnt_.topLeftCorner(2 * col_, 2 * col_); }
-  auto &wnt_raw() { return wnt_; }
-
-  auto &wn1() { return wn1_; }
-
-  auto ws() { return ws_.leftCols(col_); }
-
-  auto wy() { return wy_.leftCols(col_); }
-
-  auto ss() { return ss_.topLeftCorner(col_, col_); }
-
-  auto sy() { return sy_.topLeftCorner(col_, col_); }
-
-  auto wtt() { return wtt_.topLeftCorner(col_, col_); }
-
-  auto p() { return p_.head(2 * col_); }
-
-  auto v() { return v_.head(2 * col_); }
-
-  auto c() { return c_.head(2 * col_); }
-
-  auto wbp() { return wbp_.head(2 * col_); }
-
-  auto smul() { return smul_.head(nonnegative(col_ - 1)); }
-
-  auto &z() { return z_; }
-  auto &xcp() { return z_; }
-
-  auto &r() { return r_; }
-  auto rfree() { return r_.head(nfree_).matrix(); }
-
-  auto &t() { return t_; }
-  auto &xp() { return t_; }
-
-  auto &d() { return d_; }
-
-  auto &iwhere() { return iwhere_; }
-
-  auto free() const { return free_bound_.head(nfree_); }
-
-  auto bound() const { return free_bound_.tail(n() - nfree_); }
-
-  auto enter() const { return enter_leave_.head(nenter_); }
-
-  auto leave() const { return enter_leave_.tail(nleave_); }
-
-  auto &brks() { return brks_; }
-
-  double theta() const { return theta_; }
-
-  int col() const { return col_; }
-
-  int prev_col() const { return prev_col_; }
-
-  bool updated() const { return updated_; }
-
-  /* Test utils; normally not used */
-
-  void update_col(int col) { col_ = col; }
-
-  void update_theta(double theta) { theta_ = theta; }
-
-  auto &free_bound() { return free_bound_; }
-
-  void update_nfree(int nfree) { nfree_ = nfree; }
-
-private:
-  bool freev(int iter);
-
-  bool formk();
-
-  bool prepare_lnsrch(const ArrayXd &gx, double sbgnrm, int iter);
-
-  bool prepare_next_iter(double gd, double ginit, double step, double dtd);
-
-  double d_dot(const ArrayXd &gx) const { return gx.matrix().dot(d_.matrix()); }
-
-  void reset_memory() {
-    theta_ = 1.0;
-    col_ = prev_col_ = 0;
-    updated_ = false;
-  }
-
-  /*  System, size n */
-  MutRef<ArrayXd> x_;
-  /* Bound constraints */
-  internal::LbfgsbBounds bounds_;
-
-  /* (2m, 2m) */
-  MatrixXd wnt_, wn1_;
-  /* (n, m) */
-  MatrixXd ws_, wy_;
-  /* (m, m) */
-  MatrixXd ss_, sy_, wtt_;
-  /* 2*m */
-  VectorXd p_, v_, c_, wbp_;
-  /* n */
-  ArrayXd z_, r_, t_, d_;
-  /* m-1 */
-  ArrayXd smul_;
-
-  /* n */
-  ArrayXi iwhere_, free_bound_, enter_leave_;
-
-  internal::ClearablePQ<internal::CauchyBrkpt, std::greater<>> brks_;
-
-  double theta_;
-  int col_, prev_col_;
-  bool constrained_, boxed_, updated_;
-
-  int nfree_, nenter_, nleave_;
-};
-
-template <class FuncGrad>
-LbfgsbResult LBfgsB::minimize(FuncGrad fg, const double factr,
-                              const double pgtol, const int maxiter,
-                              const int maxls) {
-  const double tol = factr * internal::kEpsMach;
-
-  ArrayXd gx(n());
-  double fx = fg(gx, x());
-
-  internal::AllowEigenMallocScoped<false> ems;
-
-  double sbgnrm = internal::lbfgsb_projgr(x(), gx, bounds());
-  if (sbgnrm <= pgtol)
-    return { LbfgsbResultCode::kSuccess, 0, fx, std::move(gx) };
-
-  reset_memory();
-  nfree_ = n();
-  nenter_ = nleave_ = 0;
-
-  int iter = 0;
-  for (; iter < maxiter; ++iter) {
-    if (!prepare_lnsrch(gx, sbgnrm, iter)) {
-      reset_memory();
-      continue;
-    }
-
-    d() = z() - x();
-
-    double gd = d_dot(gx);
-    if (gd >= 0) {
-      if (col() == 0)
-        return { LbfgsbResultCode::kAbnormalTerm, iter + 1, fx, std::move(gx) };
-
-      reset_memory();
-      continue;
-    }
-
-    t() = x();
-    r() = gx;
-
-    internal::LbfgsbLnsrch lnsrlb(x(), t(), z(), d(), bounds(), fx, gd, iter,
-                                  constrained(), boxed());
-    bool converged = false;
-    for (int i = 0; !converged && i < maxls; ++i) {
-      NURI_EIGEN_ALLOW_MALLOC(true);
-      fx = fg(gx, x());
-      NURI_EIGEN_ALLOW_MALLOC(false);
-
-      gd = d_dot(gx);
-      converged = lnsrlb.search(fx, gd);
-    }
-
-    if (!converged) {
-      x() = t();
-      fx = lnsrlb.finit();
-      gx = r();
-
-      if (col() == 0)
-        return { LbfgsbResultCode::kAbnormalTerm, iter + 1, fx, std::move(gx) };
-
-      reset_memory();
-      continue;
-    }
-
-    sbgnrm = internal::lbfgsb_projgr(x(), gx, bounds());
-    if (sbgnrm <= pgtol)
-      return { LbfgsbResultCode::kSuccess, iter + 1, fx, std::move(gx) };
-
-    double ddum = std::max({ std::abs(lnsrlb.finit()), std::abs(fx), 1.0 });
-    if (lnsrlb.finit() - fx <= tol * ddum)
-      return { LbfgsbResultCode::kSuccess, iter + 1, fx, std::move(gx) };
-
-    r() = gx - r();
-    if (!prepare_next_iter(gd, lnsrlb.ginit(), lnsrlb.step(), lnsrlb.dtd()))
-      reset_memory();
-  }
-
-  return { LbfgsbResultCode::kMaxIterReached, maxiter, fx, std::move(gx) };
-}
 
 /**
  * @brief Minimize a function using L-BFGS-B algorithm.
@@ -517,7 +553,7 @@ LbfgsbResult LBfgsB::minimize(FuncGrad fg, const double factr,
  *         value, and final gradient.
  *
  * @note The input `x` will be modified in-place.
- * @sa LBfgsB
+ * @sa LBfgs
  *
  * References:
  * - RH Byrd, P Lu, J Nocedal and C Zhu. *SIAM J. Sci. Stat. Comput.* **1995**,
@@ -568,15 +604,164 @@ LbfgsbResult LBfgsB::minimize(FuncGrad fg, const double factr,
  * \endcode
  */
 template <class FuncGrad>
-LbfgsbResult l_bfgs_b(FuncGrad &&fg, MutRef<ArrayXd> x, const ArrayXi &nbd,
-                      const Array2Xd &bounds, const int m = 10,
-                      const double factr = 1e+7, const double pgtol = 1e-5,
-                      const int maxiter = 15000, const int maxls = 20) {
+LbfgsResult l_bfgs_b(FuncGrad &&fg, MutRef<ArrayXd> x, const ArrayXi &nbd,
+                     const Array2Xd &bounds, const int m = 10,
+                     const double factr = 1e+7, const double pgtol = 1e-5,
+                     const int maxiter = 15000, const int maxls = 20) {
   bool args_ok = internal::lbfgsb_errclb(x, nbd, bounds, m, factr);
   if (!args_ok)
-    return { LbfgsbResultCode::kInvalidInput, 0, 0, {} };
+    return { LbfgsResultCode::kInvalidInput, 0, 0, {} };
 
-  LBfgsB lbfgsb(x, { nbd, bounds }, m);
+  LBfgs<internal::LBfgsBImpl> lbfgsb(
+      x, internal::LBfgsBImpl(x, { nbd, bounds }, m), m);
+  return lbfgsb.minimize(std::forward<FuncGrad>(fg), factr, pgtol, maxiter,
+                         maxls);
+}
+
+namespace internal {
+  class LbfgsLnsrch {
+  public:
+    LbfgsLnsrch(MutRef<ArrayXd> &x, const ArrayXd &t, const ArrayXd &z,
+                const ArrayXd &d, double f0, double g0, int iter,
+                double ftol = 1e-3, double gtol = 0.9,
+                double xtol = 0.1) noexcept;
+
+    bool search(double f, double g);
+
+    double dtd() const { return dtd_; }
+
+    double step() const { return dcsrch_.step(); }
+
+    double xstep() const { return step() * dnorm_; }
+
+    double finit() const { return dcsrch_.finit(); }
+
+    double ginit() const { return dcsrch_.ginit(); }
+
+  private:
+    MutRef<ArrayXd> &x() { return *x_; }
+
+    const ArrayXd &t() const { return *t_; }
+
+    const ArrayXd &z() const { return *z_; }
+
+    const ArrayXd &d() const { return *d_; }
+
+    void step_x();
+
+    MutRef<ArrayXd> *x_;
+    const ArrayXd *t_, *z_, *d_;
+    double dtd_, dnorm_;
+
+    Dcsrch dcsrch_;
+  };
+
+  class LBfgsImpl {
+  public:
+    LBfgsImpl(): LBfgsImpl(10) { }
+
+    explicit LBfgsImpl(int m);
+
+    static LbfgsLnsrch lnsrch(MutRef<ArrayXd> &x, const ArrayXd &t,
+                              const ArrayXd &z, const ArrayXd &d, double f0,
+                              double g0, int iter, double ftol = 1e-3,
+                              double gtol = 0.9, double xtol = 0.1) noexcept;
+
+    static void reset() { }
+
+    static bool prepare_lnsrch(LBfgs<LBfgsImpl> &lbfgs, const ArrayXd &gx,
+                               double sbgnrm, int iter);
+
+    static double projgr(ConstRef<ArrayXd> x, const ArrayXd &gx);
+
+    MatrixXd &wn1() { return wn1_; }
+
+  private:
+    /* (2m, m) */
+    MatrixXd wn1_;
+  };
+
+  extern bool lbfgs_errclb(const ArrayXd &x, int m, double factr);
+}  // namespace internal
+
+/**
+ * @brief Minimize a function using L-BFGS algorithm.
+ *
+ * @tparam FuncGrad Function object that computes the function value and
+ *         gradient. Function value should be returned and gradient should be
+ *         updated in the input gradient vector.
+ * @param fg Function object.
+ * @param x Initial guess. Will be modified in-place.
+ * @param m The maximum number of variable metric corrections used to define the
+ *          limited memory matrix.
+ * @param factr Stop when function value changes by less than this factor times
+ *        the machine precision.
+ * @param pgtol Stop when the projected gradient is less than this value.
+ * @param maxiter Maximum number of iterations.
+ * @param maxls Maximum number of line search steps.
+ * @return A struct with the result code, number of iterations, final function
+ *         value, and final gradient.
+ *
+ * @note The input `x` will be modified in-place.
+ * @sa LBfgs
+ *
+ * References:
+ * - RH Byrd, P Lu, J Nocedal and C Zhu. *SIAM J. Sci. Stat. Comput.* **1995**,
+ *   *16* (5), 1190-1208. DOI:[10.1137/0916069](https://doi.org/10.1137/0916069)
+ * - C Zhu, RH Byrd, J Nocedal. *ACM Trans. Math. Softw.* **1997**, *23* (4),
+ *   550-560. DOI:[10.1145/279232.279236](https://doi.org/10.1145/279232.279236)
+ * - JL Morales, J Nocedal. *ACM Trans. Math. Softw.* **2011**, *38* (1), 7.
+ *   DOI:[10.1145/2049662.2049669](https://doi.org/10.1145/2049662.2049669)
+ *
+ * This implementation is based on the C implementation of L-BFGS-B in the SciPy
+ * library, which is a translation of the original Fortran code by Ciyou Zhu,
+ * Richard Byrd, and Jorge Noceda. Both are released under the BSD 3-Clause
+ * License and the original license is included below.
+ *
+ * \code{.unparsed}
+ * Copyright (c) 2011 Ciyou Zhu, Richard Byrd, Jorge Nocedal and Jose Luis
+ * Morales.
+ * Copyright (c) 2001-2002 Enthought, Inc. 2003-2024, SciPy Developers.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above
+ *    copyright notice, this list of conditions and the following
+ *    disclaimer in the documentation and/or other materials provided
+ *    with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * \endcode
+ */
+template <class FuncGrad>
+LbfgsResult l_bfgs(FuncGrad &&fg, MutRef<ArrayXd> x, const int m = 10,
+                   const double factr = 1e+7, const double pgtol = 1e-5,
+                   const int maxiter = 15000, const int maxls = 20) {
+  bool args_ok = internal::lbfgs_errclb(x, m, factr);
+  if (!args_ok)
+    return { LbfgsResultCode::kInvalidInput, 0, 0, {} };
+
+  LBfgs<internal::LBfgsImpl> lbfgsb(x, internal::LBfgsImpl(m), m);
   return lbfgsb.minimize(std::forward<FuncGrad>(fg), factr, pgtol, maxiter,
                          maxls);
 }
